+++ src/types.ts
@@ -23,6 +23,8 @@ type MessagePath = MessagePathItem[];
 export interface Impact {
   id: number;
   impact: string;
+  changeable?: boolean;
+  replacedImpactId: number | null;
 }
 
 export interface MessageCharacter {