+++ src/Charisma.ts
@@ -143,18 +143,6 @@ class Charisma extends EventEmitter<"ready" | "connect" | "error"> {
     return conversationId;
   }
 
-  public static async createImpactConversation(
-    token: string,
-    impactId: number,
-  ): Promise<ConversationId> {
-    const { conversationId } = await fetchJson<CreateConversationResult>(
-      `${Charisma.charismaUrl}/play/conversation/impact`,
-      { impactId },
-      { headers: { Authorization: `Bearer ${token}` } },
-    );
-    return conversationId;
-  }
-
   public static async getPlaythroughInfo(
     token: string,
   ): Promise<GetPlaythroughInfoResult> {
@@ -229,10 +217,6 @@ class Charisma extends EventEmitter<"ready" | "connect" | "error"> {
     characterId: number,
   ): Promise<ConversationId> {
     return Charisma.createCharacterConversation(this.token, characterId);
-  }
-
-  public createImpactConversation(impactId: number): Promise<ConversationId> {
-    return Charisma.createImpactConversation(this.token, impactId);
   }
 
   public getPlaythroughInfo(): Promise<GetPlaythroughInfoResult> {